requirements.txt
openai~=0.27.4
aiogram~=2.25.1
tenacity~=8.2.2
chardet
aiodns
aiohttp
charset-normalizer
regex
python-dotenv~=1.0.0